// src/app/customers/interfaces/customer.interface.ts
// Generated by https://quicktype.io

export interface Customer {
    id:                          number;
    identificationType:          number;
    identification:              string;
    dv:                          string;
    customerName:                string;
    apellido1:                   string;
    apellido2:                   string;
    nombre1:                     string;
    nombre2:                     string;
    tributaryIdentificationName: string;
    electronicMail:              string;
    telephone:                   string;
    paisId:                      number;
    municipioId:                 number;
    addressLine1:                string;
    tipoPersonaId:               number;
    tipoRegimenId:               number;
    schemeIdentifier:            string;
    responsabilidadesFiscales:   string[];
    plazoCredito:                number;
    bloqueoMora:                 number;
    cupoCredito:                 number;
    tipoPrecioId:                number;
    vendedorId:                  number;
    zonaId:                      number;
    aplicaFE:                    number;
    activo:                      number;
    isResident:                  number;
    isReteFuente:                number;
    isReteIVA:                   number;
    isReteICA:                   number;
    retefteId:                   null;
    porcreteica:                 string;
    discountParams:              DiscountParams;
}

export interface DiscountParams {
    items: Item[];
}

export interface Item {
    clientAllowanceId: null;
    days:              number;
    rate:              number;
}